add unit tests for sensors.py
